+++ packages/next-common/components/summary/polkadotTreasurySummary/ambassador/index.jsx
@@ -4,7 +4,7 @@ import FiatPriceLabel from "../common/fiatPriceLabel";
 import SummaryItem from "next-common/components/summary/layout/item";
 import { toPrecision } from "next-common/utils";
 import { SYMBOL_DECIMALS } from "next-common/utils/consts/asset";
-import { AmbassadorAccount } from "../hook/useQueryAmbassadorBalance";
+import { AmbassadorAccount } from "next-common/context/treasury/polkadotTreasury/hooks/useQueryAmbassadorBalance";
 import Link from "next/link";
 import TokenSymbolAsset from "../common/tokenSymbolAsset";
