+++ packages/next-common/components/summary/polkadotTreasurySummary/index.jsx
@@ -2,7 +2,7 @@ import TotalTreasury from "./totalTreasury";
 import RelayChainTreasury from "./relayChainTreasury";
 import TreasuryOnAssetHub from "./treasuryOnAssetHub";
 import TreasuryOnHydration from "./treasuryOnHydration";
-import { PolkadotTreasurySummaryProvider } from "./context";
+import PolkadotTreasuryProvider from "next-common/context/treasury/polkadotTreasury";
 import { TreasuryOnHydrationProvider } from "./context/treasuryOnHydration";
 import { MythTokenAssetsProvider } from "./context/mythTokenAssets";
 import FellowshipTreasuryOnAssetHub from "./fellowshipTreasuryOnAssetHub";
@@ -29,12 +29,12 @@ function PolkadotTreasurySummaryInContext() {
 
 export default function PolkadotTreasurySummary() {
   return (
-    <PolkadotTreasurySummaryProvider>
+    <PolkadotTreasuryProvider>
       <TreasuryOnHydrationProvider>
         <MythTokenAssetsProvider>
           <PolkadotTreasurySummaryInContext />
         </MythTokenAssetsProvider>
       </TreasuryOnHydrationProvider>
-    </PolkadotTreasurySummaryProvider>
+    </PolkadotTreasuryProvider>
   );
 }
